Updated Union types input sequence to match expected annotation ordering

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -4,6 +4,6 @@
 from typing import Tuple
 
 
-def sum_mixed_list(mxd_lst: Tuple[float, int]) -> float:
+def sum_mixed_list(mxd_lst: Tuple[int, float]) -> float:
     """function def"""
     return float(sum(mxd_lst))
